+++ packages/node/src/machine/index.ts
@@ -1,10 +1,4 @@
 import { appIdentityToHash } from "../ethereum/utils/app-identity";
-import {
-  AppInstance,
-  AppInstanceJson,
-  StateChannel,
-  StateChannelJSON
-} from "../models";
 
 import { Opcode, Protocol } from "./enums";
 import { InstructionExecutor } from "./instruction-executor";
@@ -34,10 +28,6 @@ import {
 export { virtualChannelKey } from "./virtual-app-key";
 export {
   appIdentityToHash,
-  AppInstance,
-  AppInstanceJson,
-  StateChannel,
-  StateChannelJSON,
   InstructionExecutor,
   Context,
   Instruction,
