+++ packages/node/src/methods/app-instance/install-virtual/operation.ts
@@ -1,6 +1,7 @@
 import { AppInstanceInfo, Node } from "@counterfactual/types";
 
-import { InstructionExecutor, StateChannel } from "../../../machine";
+import { InstructionExecutor } from "../../../machine";
+import { StateChannel } from "../../../models";
 import { Store } from "../../../store";
 import { ERRORS } from "../../errors";
