+++ packages/node/src/methods/app-instance/update-state/controller.ts
@@ -2,7 +2,8 @@ import { Node, SolidityABIEncoderV2Type } from "@counterfactual/types";
 import { INVALID_ARGUMENT } from "ethers/errors";
 import Queue from "p-queue";
 
-import { InstructionExecutor, StateChannel } from "../../../machine";
+import { InstructionExecutor } from "../../../machine";
+import { StateChannel } from "../../../models";
 import { RequestHandler } from "../../../request-handler";
 import { Store } from "../../../store";
 import { NODE_EVENTS, UpdateStateMessage } from "../../../types";
