+++ packages/node/test/machine/integration/mininode.ts
@@ -4,11 +4,8 @@ import { SigningKey } from "ethers/utils";
 import { HDNode } from "ethers/utils/hdnode";
 
 import { EthereumCommitment } from "../../../src/ethereum/types";
-import {
-  InstructionExecutor,
-  Opcode,
-  StateChannel
-} from "../../../src/machine";
+import { InstructionExecutor, Opcode } from "../../../src/machine";
+import { StateChannel } from "../../../src/models";
 
 import { getRandomHDNodes } from "./random-signing-keys";
